Changed the type of the iterator from int to ssize_t.

+++ src/kernel/syscall.c
@@ -211,12 +211,12 @@ ssize_t
 sys_write(int fildes, const void *buf, size_t nbyte)
 {
     u16 *video;
-    int i;
+    ssize_t i;
     char *s;
 
     if ( 1 == fildes && NULL != buf ) {
         video = (u16 *)0xc00b8000;
-        for ( i = 0; i < nbyte; i++ ) {
+        for ( i = 0; i < (ssize_t)nbyte; i++ ) {
             *video = 0x0f00 | (u16)((char *)buf)[i];
             //*video = 0x2f00 | (u16)*s;
             video++;
